requirements.txt
sdx-gcp==1.3.2
sdc-cryptography==1.0.0
cryptography==40.0.2
PyYAML==6.0
jsonschema~=4.17.3
Werkzeug==2.3.7
arrow==1.3.0
python-dateutil~=2.8.2